IssueNo:https://gitee.com/openharmony/notification_ans_standard/issues/I4YIZ9 Description:fix codex bug Sig: SIG_ApplicationFramework Feature or Bugfix:Bugfix Binary Source:No

+++ services/ans/src/notification_preferences.cpp
@@ -317,7 +317,7 @@ ErrCode NotificationPreferences::GetNotificationSlotsNumForBundle(
     ErrCode result = ERR_OK;
     NotificationPreferencesInfo::BundleInfo bundleInfo;
     if (preferencesInfo_.GetBundleInfo(bundleOption, bundleInfo)) {
-        num = bundleInfo.GetAllSlotsSize();
+        num = static_cast<int>(bundleInfo.GetAllSlotsSize());
     } else {
         result = ERR_ANS_PREFERENCES_NOTIFICATION_BUNDLE_NOT_EXIST;
     }
